Add some extra TS rules
